refactor: extract projects page to a component for reusability
