+++ src/app/components/ProjectsPageTemplate/NoProjectsMessage.tsx
